requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.6.0
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohttp==3.11.18
alembic>=1.14.0
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.8.1
beautifulsoup4==4.13.4
click==8.1.8
cryptography==46.0.2
deep-translator==1.11.4
edge-tts==7.2.7
fastapi==0.118.3
google-generativeai==0.8.6
gTTS==2.5.4
gunicorn==23.0.0
httpx==0.27.0
Jinja2==3.1.6
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.66
openai==1.92.2
openai-whisper==20250625
pandas==2.3.3
pillow==11.3.0
psycopg2-binary==2.9.10
pydantic==2.11.10
pydantic-settings==2.10.1
pydub==0.25.1
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.5
SQLAlchemy==2.0.41
starlette==0.48.0
uvicorn==0.30.6
yt-dlp==2025.12.8
passlib[bcrypt]
python-jose[cryptography]
argon2-cffi
